# contracts/vuelos-ecommerce-openapi.yaml
...
openapi: 3.0.0
info:
  description: >-
    Contrato del sistema de Vuelos tal como está implementado: núcleo GDS (búsqueda, retención de cupos, reservas, billetes, posventa,
    check-in, webhooks), e-commerce (identidad, catálogo, ofertas, pagos simulados, órdenes) y administración (órdenes, vuelos, usuarios,
    dashboard, observabilidad).


    Convenciones comunes: prefijo /api/v1; autenticación Bearer JWT (HS256) y `ownerId` siempre igual al `sub` verificado; errores
    `application/problem+json` con el campo `code`; las escrituras críticas exigen `Idempotency-Key` (UUID); los importes en el borde de la
    API son cadenas decimales y en el sistema son enteros en unidades menores.
  title: RAM Alliance · API de Vuelos (sistema completo)
  version: '1.0'
servers:
  - description: Producción (Railway)
    url: https://proyect-flights-rda1-plantilla-production.up.railway.app/api/v1
  - description: Desarrollo local
    url: http://localhost:3000/api/v1
...
paths:
...
  /admin/ordenes/{numero}:
    get:
      description: Incluye el contacto del comprador.
      operationId: AdminController_orden
      parameters:
        - description: Número de orden
          in: path
          name: numero
          required: true
          schema:
            example: ORD-7K3M9PQ2XA
            type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminOrdenViewDto'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetailsDto'
          description: Petición inválida
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetailsDto'
          description: No autenticado
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetailsDto'
          description: Prohibido
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetailsDto'
          description: No encontrado
      security:
        - bearer: []
      summary: Detalle completo de una orden (ADMIN)
      tags:
        - E-commerce · Back-office
...
  /clientes:
    post:
      description: 'RF-IAM-001/002: cuenta gratuita vinculada a un número LATAM Pass; se envía un correo de verificación.'
      operationId: ClientesController_registrar
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegistroClienteDto'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClienteViewDto'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetailsDto'
          description: Petición inválida
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetailsDto'
          description: No encontrado
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetailsDto'
          description: Conflicto
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetailsDto'
          description: Entidad no procesable
      summary: Registrar una cuenta
      tags:
        - E-commerce · Identidad
  /clientes/{id}:
    get:
      description: Use "me" como id para la cuenta de la sesión.
      operationId: ClientesController_obtener
      parameters:
        - description: Id del cliente o el alias "me".
          in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClienteViewDto'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetailsDto'
          description: Petición inválida
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetailsDto'
          description: No autenticado
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetailsDto'
          description: Prohibido
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetailsDto'
          description: No encontrado
      security:
        - bearer: []
      summary: Perfil del cliente
      tags:
        - E-commerce · Identidad
  /clientes/{id}/ordenes:
    get:
      description: 'RF-ORD-011: más recientes primero, paginado por cursor. Use "me" como id.'
      operationId: ClienteOrdenesController_historial
      parameters:
        - description: Id del cliente o el alias "me".
          in: path
          name: id
          required: true
          schema:
            type: string
        - in: query
          name: limit
          required: false
          schema:
            default: 20
            maximum: 50
            minimum: 1
            type: number
        - description: Cursor opaco de la página anterior.
          in: query
          name: cursor
          required: false
          schema:
            type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrdenesPaginaViewDto'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetailsDto'
          description: Petición inválida
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetailsDto'
          description: No autenticado
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetailsDto'
          description: Prohibido
      security:
        - bearer: []
      summary: Historial de órdenes del cliente
      tags:
        - E-commerce · Compra y órdenes
    post:
      description: >-
        Con el número de orden o el código de reserva y un apellido de los pasajeros. Solo se puede reclamar un viaje que aún no pertenece a
        ninguna cuenta; a partir de ahí aparece en el historial.
      operationId: ClienteOrdenesController_vincular
      parameters:
        - description: Id del cliente o el alias "me".
          in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VincularOrdenDto'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrdenViewDto'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetailsDto'
          description: Petición inválida
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetailsDto'
          description: No autenticado
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetailsDto'
          description: Prohibido
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetailsDto'
          description: No encontrado
        '429':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetailsDto'
          description: Demasiadas peticiones
      security:
        - bearer: []
      summary: Agregar a mi cuenta un viaje comprado como invitado
      tags:
        - E-commerce · Compra y órdenes
...
  /ofertas:
    post:
      description: >-
        RF-CRT-001: consolida los trayectos elegidos (ida y, si aplica, vuelta), su familia y la composición de pasajeros, fija el mercado y
        la moneda (RN-01) y retiene el inventario hasta `venceEn`. Requiere sesión: de cliente o de invitado (POST /auth/invitado).
      operationId: OfertasController_armar
      parameters:
        - in: header
          name: idempotency-key
          required: true
          schema:
            type: string
        - in: header
          name: Idempotency-Key
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ArmarOfertaDto'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OfertaViewDto'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetailsDto'
          description: Petición inválida
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetailsDto'
          description: No autenticado
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetailsDto'
          description: No encontrado
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetailsDto'
          description: Conflicto
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetailsDto'
          description: Entidad no procesable
        '429':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetailsDto'
          description: Demasiadas peticiones
        '503':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetailsDto'
          description: Servicio no disponible
      security:
        - bearer: []
      summary: Armar una oferta de viaje
      tags:
        - E-commerce · Ofertas y checkout
  /ofertas/{id}:
    delete:
      operationId: OfertasController_cancelar
      parameters:
        - in: path
          name: id
          required: true
          schema:
            format: uuid
            type: string
      responses:
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetailsDto'
          description: Petición inválida
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetailsDto'
          description: No autenticado
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetailsDto'
          description: Prohibido
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetailsDto'
          description: No encontrado
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetailsDto'
          description: Conflicto
      security:
        - bearer: []
      summary: Descartar la oferta y liberar el inventario
      tags:
        - E-commerce · Ofertas y checkout
    get:
      description: 'RF-CRT-004/006: resumen con total persistente, vigencia y lo que falta antes de pagar.'
      operationId: OfertasController_obtener
      parameters:
        - in: path
          name: id
          required: true
          schema:
            format: uuid
            type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OfertaViewDto'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetailsDto'
          description: Petición inválida
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetailsDto'
          description: No autenticado
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetailsDto'
          description: Prohibido
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetailsDto'
          description: No encontrado
      security:
        - bearer: []
      summary: Consultar una oferta vigente
      tags:
        - E-commerce · Ofertas y checkout
...
  /ofertas/{id}/compra:
    post:
      description: >-
        SRS §8.4 (pagarYEmitirOrden): revalida el precio, autoriza el pago (antifraude + pasarela), crea la orden y emite la reserva y los
        billetes en una sola transacción, y captura el cobro. Si la emisión falla después de autorizar el pago, la autorización se anula
        automáticamente (RN-19). Un reintento con la misma Idempotency-Key devuelve el resultado original y nunca cobra dos veces
        (RF-PAY-008). Un pago rechazado devuelve 402 y la oferta sigue vigente para probar otro medio (RF-PAY-009).
      operationId: ComprasController_comprar
      parameters:
        - in: path
          name: id
          required: true
          schema:
            format: uuid
            type: string
        - in: header
          name: idempotency-key
          required: true
          schema:
            type: string
        - in: header
          name: Idempotency-Key
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompraDto'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrdenViewDto'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetailsDto'
          description: Petición inválida
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetailsDto'
          description: No autenticado
        '402':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetailsDto'
          description: Error
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetailsDto'
          description: Prohibido
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetailsDto'
          description: No encontrado
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetailsDto'
          description: Conflicto
        '410':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetailsDto'
          description: Recurso expirado (Hold o Cotización)
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetailsDto'
          description: Entidad no procesable
        '429':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetailsDto'
          description: Demasiadas peticiones
        '502':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetailsDto'
          description: Error
        '503':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetailsDto'
          description: Servicio no disponible
      security:
        - bearer: []
      summary: Pagar y emitir la orden
      tags:
        - E-commerce · Compra y órdenes
...
  /ordenes:
    get:
      description: 'RF-ORD-010 / RF-CHK-001: acceso público para quien compró sin cuenta. No devuelve datos de contacto y limita los intentos por IP.'
      operationId: OrdenesController_recuperar
      parameters:
        - in: query
          name: numero
          required: false
          schema:
            example: ORD-7K3M9PQ2XA
            type: string
        - description: Código de reserva (PNR)
          in: query
          name: pnr
          required: false
          schema:
            example: ABC123
            type: string
        - description: Apellido de cualquier pasajero.
          in: query
          name: apellido
          required: true
          schema:
            example: Tapia
            type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrdenViewDto'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetailsDto'
          description: Petición inválida
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetailsDto'
          description: No encontrado
        '429':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetailsDto'
          description: Demasiadas peticiones
      summary: Recuperar un viaje con número de orden o PNR y apellido
      tags:
        - E-commerce · Compra y órdenes
  /ordenes/{numero}:
    get:
      operationId: OrdenesController_obtener
      parameters:
        - description: Número de orden
          in: path
          name: numero
          required: true
          schema:
            example: ORD-7K3M9PQ2XA
            type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrdenViewDto'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetailsDto'
          description: Petición inválida
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetailsDto'
          description: No autenticado
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetailsDto'
          description: No encontrado
      security:
        - bearer: []
      summary: Consultar una orden propia
      tags:
        - E-commerce · Compra y órdenes
...
  /tickets/verificar:
    get:
      description: >-
        Acceso público: quien escanea el QR de un pasajero ve si el billete es auténtico y está emitido, con el vuelo y el estado. La firma
        (HMAC) impide inventar o alterar códigos; no devuelve nombres ni contacto. Limita los intentos por IP.
      operationId: VerificacionBilletesController_verificar
      parameters:
        - description: Texto del código QR de un billete.
          in: query
          name: codigo
          required: true
          schema:
            example: v1.1234567890123.ABC234.abcdefghijklmnopqrstuv
            type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerificacionBilleteViewDto'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetailsDto'
          description: Petición inválida
        '429':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetailsDto'
          description: Demasiadas peticiones
      summary: Verificar un billete a partir de su código QR
      tags:
        - E-commerce · Verificación de billetes
...
components:
  schemas:
...
    AdminOrdenViewDto:
      properties:
        _links:
          additionalProperties:
            format: uri
            type: string
          description: HATEOAS links (Richardson Maturity Model Level 3)
          type: object
        bookingId:
          description: >-
            Reserva del núcleo de vuelos tras la emisión: es el `bookingId` de las rutas de equipaje, cambio, cancelación, check-in y pases.
            Solo en la vista del propietario.
          format: uuid
          type: string
        canal:
          type: string
        clienteId:
          description: Cuenta del comprador; null si compró como invitado.
          format: uuid
          nullable: true
          type: string
        comprador:
          enum:
            - cliente
            - invitado
          type: string
        contacto:
          description: Solo en la vista del propietario; la recuperación pública lo enmascara.
          type: object
        creadaEn:
          type: string
        estado:
          enum:
            - PENDIENTE_PAGO
            - PAGO_EN_VERIFICACION
            - PAGADA
            - EMITIDA
            - FALLIDA_COMPENSADA
            - MODIFICADA
            - DEVOLUCION_EN_CURSO
            - REEMBOLSADA
            - EN_VIAJE
            - COMPLETADA
            - EXPIRADA
          type: string
        historial:
          items:
            type: string
          type: array
        itinerarios:
          items:
            $ref: '#/components/schemas/ItinerarioOrdenDto'
          type: array
        mercado:
          type: string
        numeroOrden:
          example: ORD-7K3M9PQ2XA
          type: string
        ordenId:
          format: uuid
          type: string
        pago:
          type: object
        pasajeros:
          items:
            $ref: '#/components/schemas/PasajeroOrdenViewDto'
          type: array
        pnr:
          description: Localizador de la reserva (RN-20).
          nullable: true
          type: string
        total:
          $ref: '#/components/schemas/MontoDto'
      required:
        - ordenId
        - numeroOrden
        - estado
        - mercado
        - total
        - itinerarios
        - pasajeros
        - pago
        - creadaEn
        - historial
        - comprador
        - canal
      type: object
...
    ClienteViewDto:
      properties:
        _links:
          additionalProperties:
            format: uri
            type: string
          description: HATEOAS links (Richardson Maturity Model Level 3)
          type: object
        apellidos:
          type: string
        canalNotificacion:
          enum:
            - EMAIL
            - SMS
            - PUSH
            - WHATSAPP
          type: string
        clienteId:
          format: uuid
          type: string
        consentimientoMarketing:
          type: boolean
        correo:
          type: string
        correoVerificado:
          type: boolean
        idioma:
          type: string
        mercado:
          type: string
        nombres:
          type: string
        numeroSocio:
          description: Número LATAM Pass vinculado a la cuenta.
          type: string
      required:
        - clienteId
        - correo
        - correoVerificado
        - nombres
        - apellidos
        - numeroSocio
        - mercado
        - idioma
        - canalNotificacion
        - consentimientoMarketing
      type: object
...
    MontoDto:
      properties:
        moneda:
          example: USD
          type: string
        monto:
          description: Importe con los decimales de la moneda del mercado.
          example: '345.00'
          type: string
      required:
        - moneda
        - monto
      type: object
...
    OrdenViewDto:
      properties:
        _links:
          additionalProperties:
            format: uri
            type: string
          description: HATEOAS links (Richardson Maturity Model Level 3)
          type: object
        bookingId:
          description: >-
            Reserva del núcleo de vuelos tras la emisión: es el `bookingId` de las rutas de equipaje, cambio, cancelación, check-in y pases.
            Solo en la vista del propietario.
          format: uuid
          type: string
        contacto:
          description: Solo en la vista del propietario; la recuperación pública lo enmascara.
          type: object
        creadaEn:
          type: string
        estado:
          enum:
            - PENDIENTE_PAGO
            - PAGO_EN_VERIFICACION
            - PAGADA
            - EMITIDA
            - FALLIDA_COMPENSADA
            - MODIFICADA
            - DEVOLUCION_EN_CURSO
            - REEMBOLSADA
            - EN_VIAJE
            - COMPLETADA
            - EXPIRADA
          type: string
        historial:
          items:
            type: string
          type: array
        itinerarios:
          items:
            $ref: '#/components/schemas/ItinerarioOrdenDto'
          type: array
        mercado:
          type: string
        numeroOrden:
          example: ORD-7K3M9PQ2XA
          type: string
        ordenId:
          format: uuid
          type: string
        pago:
          type: object
        pasajeros:
          items:
            $ref: '#/components/schemas/PasajeroOrdenViewDto'
          type: array
        pnr:
          description: Localizador de la reserva (RN-20).
          nullable: true
          type: string
        total:
          $ref: '#/components/schemas/MontoDto'
      required:
        - ordenId
        - numeroOrden
        - estado
        - mercado
        - total
        - itinerarios
        - pasajeros
        - pago
        - creadaEn
        - historial
      type: object
    OrdenesPaginaViewDto:
      properties:
        items:
          items:
            $ref: '#/components/schemas/OrdenViewDto'
          type: array
        nextCursor:
          type: string
      required:
        - items
      type: object
...
    ProblemDetailsDto:
      properties:
        code:
          description: >-
            The contract's ProblemDetails.code enum, plus
            UNAUTHORIZED/FORBIDDEN/NOT_FOUND/CONFLICT/INTERNAL_ERROR/NOT_IMPLEMENTED/SERVICE_UNAVAILABLE for statuses the contract does not
            define.
          example: SEAT_TAKEN
          type: string
        detail:
          type: string
        invalidParams:
          items:
            $ref: '#/components/schemas/InvalidParamDto'
          type: array
        status:
          example: 409
          type: number
        title:
          type: string
        type:
          example: https://api.booking-hub.com/errors/seat-taken
          type: string
      required:
        - type
        - title
        - status
        - code
      type: object
...
    RegistroClienteDto:
      properties:
        aceptaTerminos:
          description: 'Debe ser true: aceptación de términos y política de datos.'
          type: boolean
        apellidos:
          type: string
        consentimientoMarketing:
          default: false
          description: 'Consentimiento de marketing: separado y no obligatorio.'
          type: boolean
        contrasena:
          description: Mínimo 10 caracteres, con al menos una letra y un número.
          minLength: 10
          type: string
        correo:
          format: email
          type: string
        fechaNacimiento:
          format: date
          type: string
        mercado:
          default: ec
          example: ec
          type: string
        nombres:
          type: string
        telefono:
          description: E.164
          example: '+593999999999'
          type: string
      required:
        - correo
        - contrasena
        - nombres
        - apellidos
        - fechaNacimiento
        - aceptaTerminos
      type: object
...
    VincularOrdenDto:
      properties:
        apellido:
          description: Apellido de cualquier pasajero.
          example: Tapia
          type: string
        numero:
          example: ORD-7K3M9PQ2XA
          type: string
        pnr:
          description: Código de reserva (PNR)
          example: ABC123
          type: string
      required:
        - apellido
      type: object
...
  securitySchemes:
    bearer:
      bearerFormat: JWT
      scheme: bearer
      type: http